Update libgit2 dep and don't bulid debuginfo
In debugging a segfault on win32 with gcc 4.9.2, I ended up updating the libgit2
dep as well as figuring out that the RelWithDebInfo profile causes a segfault
while Release does not! This looks to be a bug in gcc 4.9.2, but it'll be tough
to track down, so let's just stick with this for now.