Expose an API for reading raw object bytes
I'm just learning Rust by adding an implementation in Rust to
https://github.com/cgwalters/git-evtag

In order to compute the checksum I need access to the raw data.
Adding the libgit2-sys API should hopefully be uncontroversial, but I
had a choice of either adding this odd high level method to
`git2::Object`, or exposing safe bindings for the ODB.  Let me know if
you prefer the latter.