add support for checkout notifications
This introduces a `.notify()` method on `CheckoutBuilder` which makes it
possible to register a checkout notification callback.

This also introduces the `CheckoutNotificationType` type which
represents the type of checkout notification.

There are a variety of methods such as `.notify_on_conflict()` added to
`CheckoutBuilder` which allow granular toggling of which kinds of
notifications to invoke the callback on.