oid: Allow Oid to represent short oids
Add a `len` field to `Oid` to allow it to track if it's a short
oid to use as a prefix. This makes the `exists_prefix` API easier to
use, as you no longer need to track the length yourself.

BREAKING CHANGE: The `OidArray` no longer implements
`Deref<Target=[Oid]>`. To access the contained oids, use the iterator
returned by `OidArray::iter()`.

closes #293