Automate commit testing with proper stashing
Add a installation instructions for a git hook to automate testing
done on each commit.  Promotes proper handling of edge cases
related to testing code that will not be in the commit, which masks
errors that show up later as a nasty surprise.

Fully opt in, links to a gist on github.