Use libgit2 to track current position in a rebase
Instead of using an in-memory count field which will be wrong if this library
is used to re-start a rebase after it closes, we just track how many total
operations we expect to run into and run `git_rebase_next` until it tells us
that we are done.

We need to know the total number of rebase operations we expect to perform
ahead of time because a rebase operation can fail for a multitude of reasons,
but we _know_ that it will fail after there are no more ops to perform.
Unfortunately libgit2 returns a seemingly arbitrary error in the case where we
have walked off the end -- I'm getting a `config value 'merge.default' was not
found; class=Config (7)`.

Since we ned to return "true" errors to users, we need to know if we should
try `next` after the last commit.

This also adds a test that verifies that everything works as expected.