Update to libgit2 0.28.2
This includes a semver bump for all crates, as a few APIs have been
renamed in 0.28. Upstream provides `deprecated.h` aliases, if we'd like
to follow suit, but I feel it's better to just move forward.

- `giterr_*` -> `git_error_*`
- `GITERR_*` -> `GIT_ERROR_*`
- `git_otype` -> `git_object_t`
- `GIT_OBJ_*` -> `GIT_OBJECT_*`
- `git_ref_t` -> `git_reference_t`
- `GIT_REF_*` -> `GIT_REFERENCE_*`
- `git_idxentry_*` -> `git_index_entry_*`
- `GIT_IDXENTRY_*` -> `GIT_INDEX_ENTRY_*`
- `git_buf_free` -> `git_buf_dispose`

Proxy support is now builtin to `libgit2`, so `libcurl` is no longer a
dependency at all. Proxies were also the reason for `git2-curl` to
exist, which seems to be unnecessary for a while now, since `libgit2`
used `libcurl` itself since 0.23 anyway. Maybe `git2-curl` can just be
retired now, but I've updated it just for the sake of compatibility.

The minimum pkg-config libgit2 is now documented and tested for 0.28.
While it previously claimed support for 0.25.1, this was broken by
`git_config_entry`'s `include_depth` field, which was added *after* the
0.27 release branched.