Allow delegating all linking
This is necessary when an external build system (e.g. Buck) is taking
care of linking in git-sys, and we don't want the Rust compiler to emit
any linking arguments.

To do so, this adds a flag to just bail out of the build script and not
build any native crate if LIBGIT2_SYS_DELEGATE_LINKING is set in the
environment.