Accept all refspecs that implement AsRef<str> (#482)
* Accept all refspecs that implement AsRef<str>

This changes the signature of the Remote::fetch and Remote::push methods
to accept refspecs as slices of not only &str but anything that
implements AsRef<str> (and IntoCString and Clone). In this way, refspecs
can be provided, for example, by a &Vec<String>, which wasn’t possible
before.

* Accept AsRef<str> refspecs in Remote::download

This extends the signature of Remote::download to accept refspecs of type
&[Str], where Str: AsRef<str>, to bring it in line with the previous
change for Remote::fetch and Remote::push.