Add several bindings for mailmap
Add bindings for

- git_mailmap_free
- git_mailmap_from_repository
- git_mailmap_resolve_signature
- git_commit_author_with_mailmap
- git_commit_committer_with_mailmap

This is half the functions in the "mailmap" group and the two
convenience functions in the "commit" group, which is enough to make
mailmap support useful. I omitted git_mailmap_resolve because of the
slightly awkward function signature.

I also omitted

- git_mailmap_new
- git_mailmap_from_buffer
- git_mailmap_add_entry

which support in-memory manipulation of a mailmap representation: there
appears to be no way to write the representation to disk, neither
directly nor indirectly, so those functions seem a bit pointless.