Add more mailmap bindings
Add bindings for

- git_mailmap_new
- git_mailmap_from_buffer
- git_mailmap_add_entry

to support in-memory manipulation of a mailmap representation. Commit
c5fe47d (Add several bindings for mailmap, 2021-03-26) notes that these
seem silly on account of offering no way to write the representation
back to disk. However, they are easy to support and effectively enable
an out-of-tree .mailmap file.