Use EUSER instead of generic error to stop foreach
Previously foreach returned a generic generic error to signal the user
asked to stop iterating. If a user got that
error, they can't know for certain if it's because they asked to stop
iterating, or if a legitimate error occurred.

libgit2 provides EUSER as a reserved error code for this case, so we use it.

Closes 698