remote: remove unnecessary mutable borrow of FetchOptions
Fetch and download are not mutating the FetchOptions, so the mutable
borrow is unnecessary.

This also makes using the crate confusing, because some places like
RepoBuilder accept owned versions of the options. This means users
of the library have to deal with providing both mutable and immutable
versions of their FetchOptions.

With this patch, it will no longer be the case.