remote: remove unnecessary mutable borrow of FetchOptions
Fetch and download seem to be requesting a mutable borrow of
FetchOptions but it is not necessary.

These appear confusing and not consistent with how FetchOptions should
be dealt with, looking for example at RepoBuilder, where it accepts an
owned value.

With this change, people will deal with a more consistent interface.