Update to libgit2 1.4.5
This also reverts b9cbb445e0e08deb990c9bf605b0b0c45e9b462f which bumped
the submodule to 1.5.0-alpha. That probably should have been handled
differently. None of the commits in libgit2 between
1d5b9bd86dccc7347aaadf5e3ab122eed9413404..2a0d0bd19b5d13e2ab7f3780e094404828cbb9a7
(the previous submodule version and the merge base which was v1.4.0)
appear to be critical. Most of those were backported to the maint/v1.4
branch except for the 1.5-specific ones which aren't important to us.