Make url optional via a new "cred" feature gate for credential helpers
Add an enabled-by-default feature flag for credential helpers, which are
the only uses of the `url` crate. This reduces a
`default-features = false` build of git2 from 54 dependencies to 17, for
a configuration that many tools (e.g. anything that only cares about
local repositories) can happily use.

In addition to enabling it by default, make the "https" and "ssh"
features depend on "cred".